+++ src/features/operator/Webhook/ListWebhooks/index.tsx
@@ -4,12 +4,12 @@ import { Button } from 'antd';
 import classNames from 'classnames';
 import { useEffect } from 'react';
 
-import { ActionType } from '../../../../api-spec/protobuf/gen/js/tdex-daemon/v2/types_pb';
-import { useListWebhooksQuery, useRemoveWebhookMutation } from '../../operator.api';
+import { WebhookEvent } from '../../../../api-spec/protobuf/gen/js/tdex-daemon/v2/types_pb';
+import { useListWebhooksQuery, useRemoveWebhookMutation } from '../webhook.api';
 
 export const ListWebhooks = (): JSX.Element => {
   const { data: listWebhooks, error: listWebhooksError } = useListWebhooksQuery({
-    action: ActionType.ACTION_TYPE_ALL_ACTIONS,
+    event: WebhookEvent.WEBHOOK_EVENT_UNSPECIFIED,
   });
   const [removeWebhook] = useRemoveWebhookMutation();
